abstract common tests
