+++ edelta.parent/edelta.examples/tests/edelta/stdlib/examples/tests/EdeltaStdLibExamplesTest.java
@@ -1,48 +1,16 @@
 package edelta.stdlib.examples.tests;
 
-import static edelta.testutils.EdeltaTestUtils.assertFilesAreEquals;
 import static java.util.List.of;
-import static org.junit.Assert.fail;
 
-import java.io.IOException;
-import java.util.Collection;
-
-import org.junit.BeforeClass;
 import org.junit.Test;
 
-import edelta.lib.EdeltaEngine;
-import edelta.lib.EdeltaEngine.EdeltaRuntimeProvider;
+import edelta.examples.tests.AbstractEdeltaExamplesTest;
 import edelta.stdlib.examples.ChangeReferenceTypeExample;
 import edelta.stdlib.examples.ChangeReferenceTypeManualExample;
 import edelta.stdlib.examples.ChangeReferenceTypeMultipleExample;
 import edelta.stdlib.examples.ChangeToAbstractExample;
-import edelta.testutils.EdeltaTestUtils;
 
-public class EdeltaStdLibExamplesTest {
-
-	private static final String TESTDATA = "model/";
-	private static final String OUTPUT = "modified/";
-	private static final String EXPECTATIONS = "expectations/";
-
-	@BeforeClass
-	public static void clearOutput() throws IOException {
-		EdeltaTestUtils.cleanDirectoryRecursive(OUTPUT);
-	}
-
-	private EdeltaEngine setupEngine(
-			// String subdir, not used for the moment
-			Collection<String> ecoreFiles,
-			Collection<String> modelFiles,
-			EdeltaRuntimeProvider runtimeProvider
-		) {
-		var basedir = TESTDATA; // subdir is not used for the moment
-		var engine = new EdeltaEngine(runtimeProvider);
-		ecoreFiles
-			.forEach(fileName -> engine.loadEcoreFile(basedir + fileName));
-		modelFiles
-			.forEach(fileName -> engine.loadModelFile(basedir + fileName));
-		return engine;
-	}
+public class EdeltaStdLibExamplesTest extends AbstractEdeltaExamplesTest {
 
 	@Test
 	public void testChangeReferenceType() throws Exception {
@@ -114,34 +82,5 @@ public class EdeltaStdLibExamplesTest {
 			ecores,
 			models
 		);
-	}
-
-	private void executeSaveAndAssert(
-			EdeltaEngine engine,
-			String outputdir,
-			Collection<String> ecoreFiles,
-			Collection<String> modelFiles
-		) throws Exception {
-		engine.execute();
-		var output = OUTPUT + outputdir;
-		engine.save(output);
-		ecoreFiles.forEach
-			(fileName ->
-				assertGeneratedFiles(fileName, outputdir, output, fileName));
-		modelFiles.forEach
-			(fileName ->
-				assertGeneratedFiles(fileName, outputdir, output, fileName));
-	}
-
-	private void assertGeneratedFiles(String message, String subdir, String outputDir, String fileName) {
-		try {
-			assertFilesAreEquals(
-				message,
-				EXPECTATIONS + subdir + fileName,
-				outputDir + fileName);
-		} catch (IOException e) {
-			e.printStackTrace();
-			fail(e.getClass().getName() + ": " + e.getMessage());
-		}
 	}
 }
